refactor base
