+++ x/evm/txs/base/base.go
@@ -1,15 +1,11 @@
 package base
 
 import (
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/okex/exchain/app/refund"
 	bam "github.com/okex/exchain/libs/cosmos-sdk/baseapp"
 	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
 	authexported "github.com/okex/exchain/libs/cosmos-sdk/x/auth/exported"
-	"github.com/okex/exchain/x/common/analyzer"
 	"github.com/okex/exchain/x/evm/keeper"
 	"github.com/okex/exchain/x/evm/types"
-	"github.com/okex/exchain/x/evm/watcher"
 	"math/big"
 )
 
@@ -38,6 +34,15 @@ type Tx struct {
 	StateTransition types.StateTransition
 }
 
+
+func NewTx(config Config) *Tx {
+	return &Tx{
+		Ctx:    config.Ctx,
+		Keeper: config.Keeper,
+	}
+}
+
+
 // Prepare convert msg to state transition
 func (tx *Tx) Prepare(msg *types.MsgEthereumTx) (err error) {
 	tx.AnalyzeStart(bam.Txhash)
@@ -47,52 +52,10 @@ func (tx *Tx) Prepare(msg *types.MsgEthereumTx) (err error) {
 	return
 }
 
-// SaveTx since the txCount is used by the stateDB, and a simulated tx is run only on the node it's submitted to,
-// then this will cause the txCount/stateDB of the node that ran the simulated tx to be different with the
-// other nodes, causing a consensus error
-func (tx *Tx) SaveTx(msg *types.MsgEthereumTx) {
-	tx.AnalyzeStart(bam.SaveTx)
-	defer tx.AnalyzeStop(bam.SaveTx)
-
-	tx.Keeper.Watcher.SaveEthereumTx(msg, *tx.StateTransition.TxHash, uint64(tx.Keeper.TxCount))
-	// Prepare db for logs
-	tx.StateTransition.Csdb.Prepare(*tx.StateTransition.TxHash, tx.Keeper.Bhash, tx.Keeper.TxCount)
-	tx.StateTransition.Csdb.SetLogSize(tx.Keeper.LogSize)
-	tx.Keeper.TxCount++
-}
-
 func (tx *Tx) GetChainConfig() (types.ChainConfig, bool) {
 	return tx.Keeper.GetChainConfig(tx.Ctx)
 }
 
-func (tx *Tx) GetSenderAccount() authexported.Account {
-	pm := tx.Keeper.GenerateCSDBParams()
-	infCtx := tx.Ctx.WithGasMeter(sdk.NewInfiniteGasMeter())
-
-	return pm.AccountKeeper.GetAccount(infCtx, tx.StateTransition.Sender.Bytes())
-}
-
-func (tx *Tx) ResetWatcher(account authexported.Account) {
-	tx.Keeper.Watcher.Reset()
-	// delete account which is already in Watcher.batch
-	if account != nil {
-		tx.Keeper.Watcher.AddDelAccMsg(account, true)
-	}
-}
-
-func (tx *Tx) RefundFeesWatcher(account authexported.Account, coin sdk.Coins, price *big.Int) {
-	// fix account balance in watcher with refund fees
-	if account == nil || !tx.Keeper.Watcher.Enabled() {
-		return
-	}
-	gasConsumed := tx.Ctx.GasMeter().GasConsumed()
-	fixedFees := refund.CaculateRefundFees(gasConsumed, coin, price)
-	coins := account.GetCoins().Add2(fixedFees)
-	account.SetCoins(coins) //ignore err, no err will be returned in SetCoins
-
-	pm := tx.Keeper.GenerateCSDBParams()
-	pm.Watcher.SaveAccount(account, false)
-}
 
 // Transition execute evm tx
 func (tx *Tx) Transition(config types.ChainConfig) (result Result, err error) {
@@ -117,39 +80,6 @@ func (tx *Tx) DecorateResult(inResult *Result, inErr error) (result *sdk.Result,
 	return inResult.ExecResult.Result, inErr
 }
 
-func (tx *Tx) RestoreWatcherTransactionReceipt(msg *types.MsgEthereumTx) {
-	tx.Keeper.Watcher.SaveTransactionReceipt(
-		watcher.TransactionFailed,
-		msg,
-		*tx.StateTransition.TxHash,
-		uint64(tx.Keeper.TxCount-1),
-		&types.ResultData{}, tx.Ctx.GasMeter().GasConsumed())
-}
-
-func (tx *Tx) Commit(msg *types.MsgEthereumTx, result *Result) {
-	if result.InnerTxs != nil {
-		tx.Keeper.AddInnerTx(tx.StateTransition.TxHash.Hex(), result.InnerTxs)
-	}
-	if result.Erc20Contracts != nil {
-		tx.Keeper.AddContract(result.Erc20Contracts)
-	}
-
-	// update block bloom filter
-	if !tx.Ctx.IsAsync() {
-		tx.Keeper.Bloom.Or(tx.Keeper.Bloom, result.ExecResult.Bloom)
-	}
-	tx.Keeper.LogSize = tx.StateTransition.Csdb.GetLogSize()
-	tx.Keeper.Watcher.SaveTransactionReceipt(watcher.TransactionSuccess,
-		msg, *tx.StateTransition.TxHash,
-		uint64(tx.Keeper.TxCount-1), result.ResultData, tx.Ctx.GasMeter().GasConsumed())
-	if msg.Data.Recipient == nil {
-		tx.StateTransition.Csdb.IteratorCode(func(addr common.Address, c types.CacheCode) bool {
-			tx.Keeper.Watcher.SaveContractCode(addr, c.Code)
-			tx.Keeper.Watcher.SaveContractCodeByHash(c.CodeHash, c.Code)
-			return true
-		})
-	}
-}
 
 func (tx *Tx) EmitEvent(msg *types.MsgEthereumTx, result *Result) {
 	tx.Ctx.EventManager().EmitEvents(sdk.Events{
@@ -177,25 +107,28 @@ func (tx *Tx) EmitEvent(msg *types.MsgEthereumTx, result *Result) {
 	result.ExecResult.Result.Events = tx.Ctx.EventManager().Events()
 }
 
-func (tx *Tx) FinalizeWatcher(account authexported.Account, err error) {
-	if err != nil {
-		tx.ResetWatcher(account)
-		return
-	}
-	tx.Keeper.Watcher.Finalize()
-}
+// SaveTx check Tx do not transition state db
+func (t *Tx) SaveTx(msg *types.MsgEthereumTx) {}
 
-func (tx *Tx) AnalyzeStart(tag string) {
-	analyzer.StartTxLog(tag)
-}
+func (t *Tx) GetSenderAccount() authexported.Account { return nil }
 
-func (tx *Tx) AnalyzeStop(tag string) {
-	analyzer.StopTxLog(tag)
-}
+func (t *Tx) ResetWatcher(account authexported.Account) {}
 
-func NewTx(config Config) *Tx {
-	return &Tx{
-		Ctx:    config.Ctx,
-		Keeper: config.Keeper,
-	}
-}
+// RefundFeesWatcher refund the watcher, check Tx do not save state so. skip
+func (t *Tx) RefundFeesWatcher(account authexported.Account, coins sdk.Coins, price *big.Int) {}
+
+// RestoreWatcherTransactionReceipt check Tx do not need restore
+func (t *Tx) RestoreWatcherTransactionReceipt(msg *types.MsgEthereumTx) {}
+
+// Commit check Tx do not need
+func (t *Tx) Commit(msg *types.MsgEthereumTx, result *Result) {}
+
+// FinalizeWatcher check Tx do not need this
+func (t *Tx) FinalizeWatcher(account authexported.Account, err error) {}
+
+// AnalyzeStart check Tx do not analyze start
+func (t *Tx) AnalyzeStart(tag string) {}
+
+// AnalyzeStop check Tx do not analyze stop
+func (t *Tx) AnalyzeStop(tag string) {}
+
